requirements.txt
asgiref==3.8.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.2.19
django-allauth==65.7.0
django-environ==0.8.1
djangorestframework==3.15.2
gunicorn==20.1.0
psycopg2-binary==2.9.10
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
whitenoise==5.3.0
